+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 0.3.0
+
+Added `Show Previous`/`Show Next`/`Show Last` to chat view context menu
+
+
 ## Version 0.2.1
 
 - Improved `parseToolCalls` logic in `apiHandler.js`